+++ tests/custom.test.js
@@ -1,6 +1,6 @@
 const test = require('tape');
 const db = require('./db');
-const dumpster = require('../');
+const JsParseWiki = require('../');
 
 test('custom-made-tinywiki', function(t) {
   const dbName = 'tempwiki';
@@ -11,7 +11,7 @@ test('custom-made-tinywiki', function(t) {
     images: true
   };
   db.drop(dbName, 'pages', () => {
-    dumpster(obj, () => {
+    JsParseWiki(obj, () => {
       db.firstTen(dbName, docs => {
         t.equal(docs.length, 7, 'seven records');